[Refactor] 뷰 분리 (#9)
